requirements.txt
django==4.2.16
djangorestframework==3.15.2
python-decouple==3.8
djangorestframework-simplejwt==5.3.1
